+++ smdebug/core/utils.py
@@ -3,7 +3,6 @@ import bisect
 import json
 import os
 import re
-import shutil
 import socket
 from pathlib import Path
 from typing import Dict, List
@@ -295,84 +294,3 @@ def get_tb_worker():
 def remove_file_if_exists(file_path):
     if os.path.exists(file_path):
         os.remove(file_path)
-
-
-class SagemakerSimulator(object):
-    """
-    Creates an environment variable pointing to a JSON config file, and creates the config file.
-    Used for integration testing with zero-code-change.
-
-    If `disable=True`, then we still create the `out_dir` directory, but ignore the config file.
-    """
-
-    def __init__(
-        self,
-        json_config_path="/tmp/zcc_config.json",
-        tensorboard_dir="/tmp/tensorboard",
-        training_job_name="sm_job",
-        json_file_contents="{}",
-        cleanup=True,
-    ):
-        self.out_dir = DEFAULT_SAGEMAKER_OUTDIR
-        self.json_config_path = json_config_path
-        self.tb_json_config_path = DEFAULT_SAGEMAKER_TENSORBOARD_PATH
-        self.tensorboard_dir = tensorboard_dir
-        self.training_job_name = training_job_name
-        self.json_file_contents = json_file_contents
-        self.cleanup = cleanup
-
-    def __enter__(self):
-        if self.cleanup is True:
-            shutil.rmtree(self.out_dir, ignore_errors=True)
-        shutil.rmtree(self.json_config_path, ignore_errors=True)
-        tb_parent_dir = str(Path(self.tb_json_config_path).parent)
-        shutil.rmtree(tb_parent_dir, ignore_errors=True)
-
-        os.environ[CONFIG_FILE_PATH_ENV_STR] = self.json_config_path
-        os.environ[TENSORBOARD_CONFIG_FILE_PATH_ENV_STR] = self.tb_json_config_path
-        os.environ["TRAINING_JOB_NAME"] = self.training_job_name
-        with open(self.json_config_path, "w+") as my_file:
-            # We'll just use the defaults, but the file is expected to exist
-            my_file.write(self.json_file_contents)
-
-        os.makedirs(tb_parent_dir, exist_ok=True)
-        with open(self.tb_json_config_path, "w+") as my_file:
-            my_file.write(
-                f"""
-                {{
-                    "LocalPath": "{self.tensorboard_dir}"
-                }}
-                """
-            )
-
-        return self
-
-    def __exit__(self, *args):
-        # Throws errors when the writers try to close.
-        # shutil.rmtree(self.out_dir, ignore_errors=True)
-        if self.cleanup is True:
-            remove_file_if_exists(self.json_config_path)
-            remove_file_if_exists(self.tb_json_config_path)
-            if CONFIG_FILE_PATH_ENV_STR in os.environ:
-                del os.environ[CONFIG_FILE_PATH_ENV_STR]
-            if "TRAINING_JOB_NAME" in os.environ:
-                del os.environ["TRAINING_JOB_NAME"]
-            if TENSORBOARD_CONFIG_FILE_PATH_ENV_STR in os.environ:
-                del os.environ[TENSORBOARD_CONFIG_FILE_PATH_ENV_STR]
-
-
-class ScriptSimulator(object):
-    def __init__(self, out_dir="/tmp/test", tensorboard_dir=None):
-        self.out_dir = out_dir
-        self.tensorboard_dir = tensorboard_dir
-
-    def __enter__(self):
-        shutil.rmtree(self.out_dir, ignore_errors=True)
-        if self.tensorboard_dir:
-            shutil.rmtree(self.tensorboard_dir, ignore_errors=True)
-        return self
-
-    def __exit__(self, *args):
-        shutil.rmtree(self.out_dir, ignore_errors=True)
-        if self.tensorboard_dir:
-            shutil.rmtree(self.tensorboard_dir, ignore_errors=True)
